+++ packages/block-library/src/image/image.js
@@ -5,7 +5,6 @@ import {
 	ExternalLink,
 	FocalPointPicker,
 	ResizableBox,
-	ProgressBar,
 	Spinner,
 	TextareaControl,
 	TextControl,
@@ -272,7 +271,6 @@ function ContentOnlyControls( {
 export default function Image( {
 	temporaryURL,
 	isSideloading,
-	uploadProgress,
 	attributes,
 	setAttributes,
 	isSingleSelected,
@@ -1224,16 +1222,7 @@ export default function Image( {
 							...shadowProps.style,
 						} }
 					/>
-					{ isUploading &&
-						( uploadProgress !== undefined &&
-						uploadProgress < 100 ? (
-							<ProgressBar
-								className="wp-block-image__progress-bar"
-								value={ uploadProgress }
-							/>
-						) : (
-							<Spinner />
-						) ) }
+					{ isUploading && <Spinner /> }
 				</>
 			) }
 		</ImageWrapper>